// lib/dxp/types/dxp-api.d.ts
export interface components {
    schemas: {
// ...
        BcicLeadCreateRequest: {
            addresses: components["schemas"]["BcicLeadCreateRequestAddress"][];
            agencies?: components["schemas"]["LeadRequestAgency"][];
            chats?: components["schemas"]["LeadCreateRequestChat"][];
            /** @example BB */
            citizenshipCd: string;
            /**
             * Format: date
             * @example 1985-07-13
             */
            dateOfBirth: string;
            emails?: components["schemas"]["LeadCreateRequestEmail"][];
            employer?: string;
            /** @example EMP_FT */
            employmentStatusCd: string;
            /** @example John */
            firstName: string;
            /**
             * @description use 'GenderCd' lookup code
             * @example male
             */
            gender?: string;
            /** @example NID */
            identificationTypeCd: string;
            /** @example Jones */
            lastName: string;
            /** @example Steve */
            middleName?: string;
            /** @example ITSpecTech */
            occupationCd: string;
            /** @example GOP */
            organizationName?: string;
            phones?: components["schemas"]["LeadCreateRequestPhone"][];
            /** @example false */
            seniorPublicOfficeInd?: boolean;
            socialNets?: components["schemas"]["LeadCreateRequestSocialNet"][];
            /**
             * @description use 'BcicSourceOfFund' lookup code
             * @example Salary
             */
            sourceOfFundCd: string;
            /** @example Investments */
            sourceOfFundDetails?: string;
            /** @example 2AAA */
            taxId: string;
            /** @example Party President */
            titlePEP?: string;
            /** @example 123456789 */
            trn?: string;
            webAddresses?: components["schemas"]["LeadCreateRequestWebAddress"][];
        };
// ...
        BcicPolicyRateResponse: {
            additionalStampDuty?: number;
            annualPremium?: number;
            bundleDiscountAnnualPremium?: number;
            currencyCode?: string;
            deposit?: number;
            discount?: number;
            enedorsementApRp?: number;
            fees?: number;
            financeCharge?: number;
            fullTermPremium?: number;
            installments?: components["schemas"]["InstallmentDTO"][];
            interestRate?: number;
            monthlyPremium?: number;
            planCd?: string;
            policyNumber?: string;
            premiumChange?: number;
            taxes?: number;
            termPremium?: number;
            totalDue?: number;
            totalPremium?: number;
            totalStampDuty?: number;
            /** Format: date-time */
            transactionEffectiveDate?: string;
        };
// ...
        Email: {
            /** @example email@email.com */
            email: string;
        };
// ...
        LeadCreationResult: {
            customerNumber?: string;
        };
// ...
        PaymentPlanSummary: {
            /** @example Once per year */
            description?: string;
            /** Format: date-time */
            effectiveDate?: string;
            /** @example true */
            enabledForNewBusiness?: boolean;
            /** @example true */
            enabledForRenewal?: boolean;
            /** Format: date-time */
            expirationDate?: string;
            /** @example depositPercentOfPremium */
            firstInstallment?: string;
            geography?: string[];
            /**
             * Format: int32
             * @example 1
             */
            installmentFrequency?: number;
            /** @example 16.66 */
            newBusinessPercent?: number;
            /**
             * Format: int32
             * @example 10
             */
            numberInstallments?: number;
            /** @example annual */
            planCd?: string;
            products?: string[];
            /** @example both */
            recurringPaymentsRule?: string;
            /** @example 9.09 */
            renewalPercent?: number;
        };
// ...
        PrecAuQuoteCreateRequest: {
            /** @example false */
            activePolicyWithBCIC?: boolean;
            readonly assignments?: components["schemas"]["PrecAuAssignment"][];
            /**
             * @description use 'BcicBranchCode' lookup code
             * @example CROSS_ROADS
             */
            branchCd?: string;
            /**
             * @description use 'ContractTermType' lookup code
             * @example 12
             */
            contractTermTypeCd: string;
            /**
             * @description use 'Country' lookup code
             * @example BB
             */
            countryCd: string;
            /**
             * @description use 'CoverageType' lookup code
             * @example Comprehensive
             */
            coverageTypeCd: string;
            /**
             * @description use 'Currency' lookup code
             * @example BBD
             */
            currencyCode: string;
            /** @example 510000 */
            customerNumber: string;
            /** Format: date-time */
            effectiveDate: string;
            /** Format: date-time */
            readonly expirationDate?: string;
            insuredsAndDrivers: components["schemas"]["PrecAuQuoteInsuredAndDriverCreateRequest"][];
            /**
             * @description use 'LOB' lookup code
             * @example AUTO
             */
            readonly lobCd?: string;
            /** Format: int64 */
            readonly pendingRevisionNumber?: number;
            readonly policyNumber?: string;
            /**
             * @description use 'PolicyStatus' lookup code
             * @example dataGather
             */
            readonly policyStatusCd?: string;
            preConversionPolicyNumber?: string;
            preConversionPolicyPremium?: number;
            /** @example false */
            premiumFinancing: boolean;
            /**
             * @description use 'BcicFinancingCompany' lookup code
             * @example BCIC
             */
            premiumFinancingCompany?: string;
            /**
             * @description use 'Product' lookup code
             * @example PREC-AU
             */
            readonly productCode?: string;
            /** Format: int64 */
            readonly productVersion?: number;
            /** Format: int64 */
            readonly revisionNumber?: number;
            /**
             * @description use 'StateProvCd' lookup code
             * @example BB-01
             */
            riskStateCd: string;
            /** @example America/Barbados */
            timeZone?: string;
            /**
             * @description use 'PolicyStatus' lookup code
             * @example dataGather
             */
            readonly timedPolicyStatusCd?: string;
            readonly totalPremium?: number;
            /** Format: date-time */
            readonly transactionEffectiveDate?: string;
            /**
             * @description use 'PolicyTransactionType' lookup code
             * @example quote
             */
            readonly transactionTypeCd?: string;
            vehicles: components["schemas"]["PrecAuVehicleCreateRequest"][];
        };
        PrecAuQuoteCreateUpdateResponse: {
            policyNumber?: string;
            /** Format: date-time */
            transactionEffectiveDate?: string;
        };
// ...
        PrecAuQuoteInsuredAndDriverCreateRequest: {
            address: components["schemas"]["PrecAuAddressCreateRequest"];
            /**
             * Format: date
             * @example 1990-12-31
             */
            dateOfBirth: string;
            driver: components["schemas"]["PrecAuQuoteDriverCreateRequest"];
            emails?: components["schemas"]["PrecAuEmailCreateRequest"][];
            /** @example ABC Company */
            employer?: string;
            /**
             * @description use 'AutoEmploymentStatusCd' lookup code
             * @example EMP_SLF
             */
            employmentStatusCd: string;
            /** @example John */
            firstName: string;
            /**
             * @description use 'GenderCd' lookup code
             * @example male
             */
            genderCd: string;
            /** @example NID */
            identificationTypeCd: string;
            insured: components["schemas"]["PrecAuQuoteInsured"];
            /** @example Jones */
            lastName: string;
            /**
             * @description use 'MaritalStatusCd' lookup code
             * @example M
             */
            maritalStatusCd: string;
            /** @example Steve */
            middleName?: string;
            /**
             * @description use 'Occupation' lookup code
             * @example Bearer
             */
            occupationCd: string;
            phones?: components["schemas"]["PrecAuPhoneCreateRequest"][];
            /** @example 2AAA */
            taxId: string;
            /** @example 123456789 */
            trn?: string;
        };
// ...
        PurchaseRequest: {
            billingAccount: components["schemas"]["QuickQuoteBillingAccountPurchase"];
            defaultPaymentMethod: components["schemas"]["QuickQuoteBillingAccountPaymentPlan"];
            paymentPlanOption: components["schemas"]["QuickQuoteBillingPaymentPlanOption"];
        };
        PurchaseResponse: {
            policyNumber?: string;
            /** Format: date-time */
            transactionEffectiveDate?: string;
        };
// ...
    };
// ...
}